ini sudah read data tanaman dan cuaca

+++ app/src/main/java/com/example/nutplant/model/DataPerangkat.java
@@ -19,7 +19,7 @@ public class DataPerangkat implements Parcelable {
     private Integer kelembabanTanah;
     @SerializedName("ph")
     @Expose
-    private Integer ph;
+    private Double ph;
     @SerializedName("kelembabanUdara")
     @Expose
     private Integer kelembabanUdara;
@@ -28,7 +28,7 @@ public class DataPerangkat implements Parcelable {
     private Integer suhuUdara;
     @SerializedName("kondisi")
     @Expose
-    private Integer kondisi;
+    private Double kondisi;
 
     public String getId() {
         return id;
@@ -54,11 +54,11 @@ public class DataPerangkat implements Parcelable {
         this.kelembabanTanah = kelembabanTanah;
     }
 
-    public Integer getPh() {
+    public Double getPh() {
         return ph;
     }
 
-    public void setPh(Integer ph) {
+    public void setPh(Double ph) {
         this.ph = ph;
     }
 
@@ -78,11 +78,11 @@ public class DataPerangkat implements Parcelable {
         this.suhuUdara = suhuUdara;
     }
 
-    public Integer getKondisi() {
+    public Double getKondisi() {
         return kondisi;
     }
 
-    public void setKondisi(Integer kondisi) {
+    public void setKondisi(Double kondisi) {
         this.kondisi = kondisi;
     }
 
@@ -109,10 +109,10 @@ public class DataPerangkat implements Parcelable {
         this.id = in.readString();
         this.tanggal = in.readString();
         this.kelembabanTanah = (Integer) in.readValue(Integer.class.getClassLoader());
-        this.ph = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.ph = (Double) in.readValue(Double.class.getClassLoader());
         this.kelembabanUdara = (Integer) in.readValue(Integer.class.getClassLoader());
         this.suhuUdara = (Integer) in.readValue(Integer.class.getClassLoader());
-        this.kondisi = (Integer) in.readValue(Integer.class.getClassLoader());
+        this.kondisi = (Double) in.readValue(Double.class.getClassLoader());
     }
 
     public static final Parcelable.Creator<DataPerangkat> CREATOR = new Parcelable.Creator<DataPerangkat>() {
